new script for loading the data base
