--- inventory/management/commands/rawmaterials.json
@@ -1,1 +1,0 @@
-[{"name":"Bread","units":50,"exp_date":"2025-09-10"},{"name":"Croissant","units":30,"exp_date":"2025-09-05"},{"name":"Muffin","units":75,"exp_date":"2025-09-20"},{"name":"Baguette","units":20,"exp_date":"2025-09-02"},{"name":"Cake","units":10,"exp_date":"2025-09-15"},{"name":"Donut","units":40,"exp_date":"2025-09-08"}]
