--- products/management/commands/products.json
@@ -1,1 +1,0 @@
-[{"name":"1014 - CAPUCCINO","description":null,"price":4000,"quantity":null,"picture":null},{"name":"1011 - SANDUCHE QUESOS","description":null,"price":13000,"quantity":null,"picture":null},{"name":"1010 - SANDUCHE PAVO","description":null,"price":21000,"quantity":null,"picture":null},{"name":"1013 - BEBIDA GRANDE","description":null,"price":4000,"quantity":null,"picture":null},{"name":"1016 - CAFE CON LECHE","description":null,"price":3000,"quantity":null,"picture":null},{"name":"1017 - MILO","description":null,"price":4000,"quantity":null,"picture":null},{"name":"1015 - TINTO","description":null,"price":2000,"quantity":null,"picture":null},{"name":"1012 - BEBIDA PEQUENIA","description":null,"price":3000,"quantity":null,"picture":null},{"name":"1008 - SANDUCHE ITALIANO","description":null,"price":20000,"quantity":null,"picture":null},{"name":"1003 - PAN SEMILLAS","description":null,"price":20000,"quantity":null,"picture":null},{"name":"1006 - TAJADA TRIGO","description":null,"price":2000,"quantity":null,"picture":null},{"name":"1009 - SANDUCHE YORK","description":null,"price":19000,"quantity":null,"picture":null},{"name":"1007 - SANDUCHE BANETON","description":null,"price":16000,"quantity":null,"picture":null},{"name":"1004 - TAJADA CHOCOLATE","description":null,"price":3000,"quantity":null,"picture":null},{"name":"1005 - TAJADA SEMILLAS","description":null,"price":3000,"quantity":null,"picture":null},{"name":"1002 - PAN CHOCOLATE","description":null,"price":20000,"quantity":null,"picture":null},{"name":"1001 - PAN CAMPESINO","description":null,"price":16000,"quantity":null,"picture":null}]
